+++ src/components/article/ProductDetails/productDetails.js
@@ -1,0 +1,2 @@
+import './productDetails.html.twig';
+import './productDetails.scss';